ev3dev-jessie: manually copy some files from base-files
This fixes the ev3dev logo missing from the motd when logging in via ssh.

Since the docker bootstrap is no longer using multistrap, we have to
install the official debian base-files package first. When upgrading
the base-files package to the ev3dev version, some of the config files
are not updated. This manually updates them.