ev3dev-buster: set nice and use PAM in console runner service
Setting nice makes user programs run with higher priority, which is
helpful when you have a tight control loop. Also having a proper PAM
session fixes some permission problems for things that use PAM for
authentication. It also has the side effect of setting the owner of
the tty so we no longer have to do that explicitly.